dung-refactor the [Role] for a more granular Role-based access control in Spring security. Each account now only has one [Role], [ROLE_ADMIN] will have implicit meanings, if an account does have multiple roles, then it means that the user of that account can use multiple services of the system, like [Manage Account] or [Place an order on item]
