+++ back-end/src/main/java/com/dnth_underdog_241/online_fashion_shopping/exception/UserNotFoundException.java
@@ -6,6 +6,6 @@ public class UserNotFoundException
 {
     public UserNotFoundException(Long id)
     {
-        super("Customer not found with id " + id);
+        super("User not found with id " + id);
     }
 }
