+++ back-end/src/main/java/com/dnth_underdog_241/online_fashion_shopping/model/user/Role.java
@@ -42,6 +42,7 @@ public class Role implements GrantedAuthority
         return getId() != null && Objects.equals(getId(), role.getId());
     }
 
+
     @Override
     public final int hashCode()
     {
